abstract case-dragging to Draggable
