+++ package.json
@@ -45,7 +45,6 @@
     "@types/prop-types": "^15.5.2",
     "@types/react": "^16.0.27",
     "@types/react-addons-shallow-compare": "^0.14.19",
-    "@types/react-beautiful-dnd": "^2.3.2",
     "@types/react-dnd": "^2.0.34",
     "@types/react-dnd-touch-backend": "^0.3.2",
     "@types/react-dom": "^16.0.3",
@@ -87,6 +86,7 @@
     "raf": "^3.4.0",
     "react": "^16.2.0",
     "react-addons-shallow-compare": "^15.5.1",
+    "react-dnd": "^2.5.4",
     "react-dnd-touch-backend": "^0.3.17",
     "react-dom": "^16.2.0",
     "react-flip-move": "^2.10.0",
@@ -116,7 +116,6 @@
     "webpack": "^2.7.0",
     "webpack-bundle-analyzer": "^2.9.1",
     "webpack-dev-server": "^2.5.1",
-    "webpack-merge": "^4.1.0",
-    "react-dnd": "^2.5.4"
+    "webpack-merge": "^4.1.0"
   }
 }
